requirements.txt
Flask
flask-login
pymongo
python-dotenv
streamlit
SpeechRecognition
transformers
scikit-learn
numpy
pydub
Flask-CORS